Refactor `RenderableText` pt. 1
